terms部分を別の階層に分離。export defaultの関係
