+++ src/rmmz/system/core/terms/index.ts
@@ -1,6 +1,4 @@
-export * from "./basic";
-export * from "./commands";
-export * from "./messages";
+export * from "./core/types";
 export * from "./types";
 export { SCHEMA_SYSTEM_TERMS_BUNDLE } from "./schema";
 export { getParamNames, makeParamNamesArray } from "./terms";
