+++ src/rmmz/system/core/terms/types.ts
@@ -1,10 +1,11 @@
-import type { Terms_Basic, Terms_BasicArray } from "./basic";
-import type { Terms_GameCommands, Terms_CommandArray } from "./commands";
-import type { Terms_Messages } from "./messages";
+import type { Terms_Basic } from "./core/basic";
+import type { Terms_GameCommands } from "./core/commands";
+import type { Terms_Messages } from "./core/messages";
 import type {
   SystemLabels_TermsParamNames,
   Terms_ParamNamesArray,
-} from "./paramArray";
+} from "./core/paramArray";
+import type { System_Terms } from "./core/types";
 
 export interface System_TextBundle extends SystemLabels_TermsParamNames {
   terms: System_Terms;
@@ -16,13 +17,6 @@ export interface System_Terms_Params {
   terms: {
     params: Terms_ParamNamesArray;
   };
-}
-
-export interface System_Terms {
-  messages: Terms_Messages;
-  commands: Terms_CommandArray;
-  params: Terms_ParamNamesArray;
-  basic: Terms_BasicArray;
 }
 
 export interface System_TermsPartial {
